Update macOS app settings and dependencies
This commit updates the macOS app settings and dependencies. Specifically, it updates the Podfile checksum, sets the CFBundleDisplayName to "VPN Line", and sets the LSApplicationCategoryType to "public.app-category.lifestyle". It also updates the entitlements file to set the aps-environment to "production" and adds security settings for network access and just-in-time compilation.